chore: done moving all the component the sub component and images to asset
