+++ src/App.css
@@ -4,7 +4,6 @@
   padding: 0;
 }
 body{
-  border: 1px solid darkcyan;
   box-sizing: border-box;
   font-family: 'Mukta', cursive;
 }
