+++ src/App.js
@@ -1,9 +1,9 @@
 import './App.css';
-
+import Dashboard from './pages/dashboardPage/Dashboard'
 function App() {
   return (
     <div className="App">
-
+      <Dashboard />
     </div>
   );
 }
