+++ src/components/dashBoard/AttendanceItems.css
@@ -5,8 +5,6 @@
 }
 .attendance-heading {
     width: 90%;
-    border-bottom: 2px solid rgb(216, 206, 206);
-    border-radius: 5px;
     margin: auto;
     padding-top: 2rem;
 }
