+++ src/components/dashBoard/AttendanceSlider.css
