+++ src/components/dashBoard/SliderBody.css
