+++ src/components/dashBoard/SliderBody.jsx
