+++ src/components/dashBoard/SliderCard.css
